fixing bugs and refactoring code

+++ frontend/admin/src/Pages/AlterItems/AlterItems.jsx
@@ -1,9 +1,12 @@
 import React, { useState, useEffect } from 'react'
+import './AlterItems.css'
+
 import Header from '../../components/Header/Header'
 import Title from '../../components/Title/Title'
+
 import { Link } from 'react-router-dom'
+
 import axios from 'axios'
-import './AlterItems.css'
 
 export default function AlterItems() {
     const [items, setItems] = useState([])
@@ -22,40 +25,46 @@ export default function AlterItems() {
 
     const handleDelete = async (item_id) => {
         try {
-            await axios.delete(`http://localhost:3333/management-item/delete-item/${item_id}`)
-                .then(async () => {
-                    await fetchAllItems()
-                })
-            // window.location.reload()
+            const confirmDelete = window.confirm('Tem certeza que deseja excluir este item?');
+            if (!confirmDelete) {
+                return;
+            }
 
+            await axios.delete(`http://localhost:3333/management-item/delete-item/${item_id}`);
+            await fetchAllItems();
             console.log(item_id);
         } catch (error) {
-            console.log(error);
+            console.log('Erro ao excluir item:', error);
         }
     }
 
     return (
         <div className='alter-items-container-all'>
+
             <Header />
 
             <div className="alter-items-container">
-                
-            <Title title="Alter Items" />
 
-                {items.map(item => (
-                    <div className="cards" key={item.id_item}>
-                        <h2>{item.item_name}</h2>
-                        <p>{item.item_description}</p>
-                        <h3>R$ {item.item_price}</h3>
+                <Title title="Alter Items" />
 
-                        <button>
-                            <Link className='link' to={`/update-item/${item.id_item}`}>Update</Link>
-                        </button>
+                <div className="alter-items-cards">
 
-                        <button onClick={() => handleDelete(item.id_item)}>Delete</button>
+                    {items.map(item => (
+                        <div className="cards" key={item.item_id}>
+                            <h2>{item.item_name}</h2>
+                            <p>{item.item_description}</p>
+                            <h3>R$ {item.item_price}</h3>
 
-                    </div>
-                ))}
+                            <button>
+                                <Link className='link' to={`/update-item/${item.item_id}`}>Update</Link>
+                            </button>
+
+                            <button onClick={() => handleDelete(item.item_id)}>Delete</button>
+
+                        </div>
+                    ))}
+
+                </div>
             </div>
         </div>
     )
